requirements.txt
openai
py-readability-metrics
PyPDF2